Uicc: Fix NullPointerException issue after turn on airplane mode
Response may be cleared when SimRecord is reset,so check if it is
null before using response.

CRs-Fixed: 521604
Change-Id: Ia2f36ab7421644de27f31e342c242f007a0d4912

Fix issue with deletion of contacts with "=".

Fix issue for deletion of contacts with "=" in name,
as it would return 3 strings instead of 2 as expected.
Use overloaded split method to return 2 strings instead
to resolve the issue.
Change-Id: Ia233ed535199c64dc2252aba8470e59aeb5694f7
CRs-Fixed: 753103