Fix the incorrect clearance of Adn records cache.
Presently if loading of a particular adn record fails then
the previous adn record cache is getting cleared, i.e
when the loading of EF_SDN fails the EF_FDN list
is getting cleared.
Fix is not to clear the previous adn record cache if loading
fails for the present adn record, but instead set the adn record
to null.

Change-Id: I24c1f47e2b9f64f30e074c34b3f192209791d256
CRs-Fixed: 784347