Fix data call retry issue on PS service loss
When data connection is in Retrying state and there is a PS Service
loss, DC is continuing in Retrying state as mAutoAttachCreation is
enabled. Due to this, irrespective of PS attach status, data call is
retried on Retry alarm expiry. Later when PS is attached, disconnect event
will be sent to cleanup existing connections if icc records are changed.
This is leading to unexpected data call tear down.

Add a check for PS service in DcRetryingState of DataConnection, and
move DC to Inactive state if PS service is lost, so that data call will
be retried only after PS attach is done.

Change-Id: Ic7c401368b91967051ff258fb6d5cf01dca088b8
CRs-Fixed: 843141