Add a flag to control SPN display
If the new added flag config_spn_display_control is set to true,
the SPN will not be catenated to the carrier text.

By default this flag is set to false to work as the same as
android original design. It could be enabled in data package
overlay if needed.

Change-Id: I74b8c18983b08cdb1d2cf7c6128e973927ec4f32